+++ src/components/visualizations/chart/highChartsCreators.ts
@@ -53,30 +53,3 @@ export function getHighchartsOptions(chartOptions: IChartOptions, drillConfig: a
     );
 }
 
-export function isDataOfReasonableSize(chartData: any, limits: IChartLimits, isViewByTwoAttributes = false) {
-    let result = true;
-
-    const seriesLimit = get(limits, "series");
-    if (seriesLimit !== undefined) {
-        result = result && chartData.series.length <= seriesLimit;
-    }
-
-    const categoriesLimit = get(limits, "categories");
-    if (categoriesLimit !== undefined) {
-        if (isViewByTwoAttributes) {
-            const categoriesLength = chartData.categories.reduce((result: number, category: any) => {
-                return result + category.categories.length;
-            }, 0);
-            result = result && categoriesLength <= categoriesLimit;
-        } else {
-            result = result && chartData.categories.length <= categoriesLimit;
-        }
-    }
-
-    const dataPointsLimit = get(limits, "dataPoints");
-    if (dataPointsLimit !== undefined) {
-        result = result && chartData.series.every((serie: any) => serie.data.length <= dataPointsLimit);
-    }
-
-    return result;
-}
