+++ src/components/visualizations/utils/chart.ts
@@ -1,5 +1,6 @@
 // (C) 2019 GoodData Corporation
 import {AFM, Execution} from "@gooddata/typings";
+import {IChartLimits} from "../../../interfaces/Config";
 import get = require("lodash/get");
 
 export function isDerivedMeasure(measureItem: Execution.IMeasureHeaderItem, afm: AFM.IAfm) {
@@ -31,4 +32,32 @@ export function findParentMeasureIndex(afm: AFM.IAfm, measureItemIndex: number):
     }
 
     return -1;
+}
+
+export function isDataOfReasonableSize(chartData: any, limits: IChartLimits, isViewByTwoAttributes = false) {
+    let result = true;
+
+    const seriesLimit = get(limits, "series");
+    if (seriesLimit !== undefined) {
+        result = result && chartData.series.length <= seriesLimit;
+    }
+
+    const categoriesLimit = get(limits, "categories");
+    if (categoriesLimit !== undefined) {
+        if (isViewByTwoAttributes) {
+            const categoriesLength = chartData.categories.reduce((result: number, category: any) => {
+                return result + category.categories.length;
+            }, 0);
+            result = result && categoriesLength <= categoriesLimit;
+        } else {
+            result = result && chartData.categories.length <= categoriesLimit;
+        }
+    }
+
+    const dataPointsLimit = get(limits, "dataPoints");
+    if (dataPointsLimit !== undefined) {
+        result = result && chartData.series.every((serie: any) => serie.data.length <= dataPointsLimit);
+    }
+
+    return result;
 }
